Move xml tests to integration tests
