+++ tests/Integration/IssueCategoryXmlTest.php
@@ -1,70 +1,63 @@
 <?php
 
-namespace Redmine\Tests\Unit;
+namespace Redmine\Tests\Integration;
 
 use DOMDocument;
 use Exception;
 use PHPUnit\Framework\TestCase;
-use Redmine\Tests\Fixtures\MockClient as TestClient;
+use Redmine\Tests\Fixtures\MockClient;
 use SimpleXMLElement;
 
-class UserXmlTest extends TestCase
+class IssueCategoryXmlTest extends TestCase
 {
     /**
-     * @var TestClient
+     * @var MockClient
      */
     private $client;
 
     public function setup(): void
     {
-        $this->client = new TestClient('http://test.local', 'asdf');
+        $this->client = new MockClient('http://test.local', 'asdf');
     }
 
     public function testCreateBlank()
     {
-        $api = $this->client->getApi('user');
-        $this->assertInstanceOf('Redmine\Api\User', $api);
+        $api = $this->client->getApi('issue_category');
+        $this->assertInstanceOf('Redmine\Api\IssueCategory', $api);
 
         $this->expectException(Exception::class);
         $this->expectExceptionMessage('Missing mandatory parameters');
 
-        $api->create();
+        $api->create('aProject');
     }
 
     public function testCreateComplex()
     {
-        $api = $this->client->getApi('user');
-        $res = $api->create([
-            'login' => 'test',
-            'firstname' => 'test',
-            'lastname' => 'test',
-            'mail' => 'test@example.com',
+        $api = $this->client->getApi('issue_category');
+        $res = $api->create('otherProject', [
+            'name' => 'test category',
         ]);
         $res = json_decode($res, true);
 
         $xml = '<?xml version="1.0"?>
-<user>
-    <login>test</login>
-    <lastname>test</lastname>
-    <firstname>test</firstname>
-    <mail>test@example.com</mail>
-</user>';
+<issue_category>
+    <name>test category</name>
+</issue_category>';
         $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
     }
 
     public function testUpdate()
     {
-        $api = $this->client->getApi('user');
+        $api = $this->client->getApi('issue_category');
         $res = $api->update(1, [
-            'firstname' => 'Raul',
+            'name' => 'new category name',
         ]);
         $res = json_decode($res, true);
 
         $xml = '<?xml version="1.0"?>
-<user>
-    <id>1</id>
-    <firstname>Raul</firstname>
-</user>';
+<issue_category>
+    <name>new category name</name>
+</issue_category>';
         $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
     }
