+++ tests/Integration/MembershipXmlTest.php
@@ -1,29 +1,29 @@
 <?php
 
-namespace Redmine\Tests\Unit;
+namespace Redmine\Tests\Integration;
 
 use DOMDocument;
 use Exception;
 use PHPUnit\Framework\TestCase;
-use Redmine\Tests\Fixtures\MockClient as TestClient;
+use Redmine\Tests\Fixtures\MockClient;
 use SimpleXMLElement;
 
-class IssueCategoryXmlTest extends TestCase
+class MembershipXmlTest extends TestCase
 {
     /**
-     * @var TestClient
+     * @var MockClient
      */
     private $client;
 
     public function setup(): void
     {
-        $this->client = new TestClient('http://test.local', 'asdf');
+        $this->client = new MockClient('http://test.local', 'asdf');
     }
 
     public function testCreateBlank()
     {
-        $api = $this->client->getApi('issue_category');
-        $this->assertInstanceOf('Redmine\Api\IssueCategory', $api);
+        $api = $this->client->getApi('membership');
+        $this->assertInstanceOf('Redmine\Api\Membership', $api);
 
         $this->expectException(Exception::class);
         $this->expectExceptionMessage('Missing mandatory parameters');
@@ -33,31 +33,39 @@ class IssueCategoryXmlTest extends TestCase
 
     public function testCreateComplex()
     {
-        $api = $this->client->getApi('issue_category');
+        $api = $this->client->getApi('membership');
         $res = $api->create('otherProject', [
-            'name' => 'test category',
+            'user_id' => 1,
+            'role_ids' => [1, 2],
         ]);
         $res = json_decode($res, true);
 
         $xml = '<?xml version="1.0"?>
-<issue_category>
-    <name>test category</name>
-</issue_category>';
+<membership>
+    <user_id>1</user_id>
+    <role_ids type="array">
+        <role_id>1</role_id>
+        <role_id>2</role_id>
+    </role_ids>
+</membership>';
         $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
     }
 
     public function testUpdate()
     {
-        $api = $this->client->getApi('issue_category');
+        $api = $this->client->getApi('membership');
         $res = $api->update(1, [
-            'name' => 'new category name',
+            'role_ids' => [1, 2],
         ]);
         $res = json_decode($res, true);
 
         $xml = '<?xml version="1.0"?>
-<issue_category>
-    <name>new category name</name>
-</issue_category>';
+<membership>
+    <role_ids type="array">
+        <role_id>1</role_id>
+        <role_id>2</role_id>
+    </role_ids>
+</membership>';
         $this->assertEquals($this->formatXml($xml), $this->formatXml($res['data']));
     }
